netwerk/cookie/test/mochitest/test_cookie_sent_upgrade_XHR.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>Upgrade-Insecure-Requests Cookie Test</title>
  <meta http-equiv="Content-Security-Policy" content="upgrade-insecure-requests">
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <script class="testbody" type="text/javascript">
    SimpleTest.waitForExplicitFinish();

    document.cookie = "testcookie=value123; path=/";

    const xhr = new XMLHttpRequest();
    // eslint-disable-next-line @microsoft/sdl/no-insecure-url
    xhr.open("GET", "http://example.com/tests/netwerk/cookie/test/mochitest/cookieUpgrade.sjs", true);

    xhr.onload = function() {
      ok(xhr.responseText.includes("scheme=https"), "Request was upgraded to HTTPS");
      ok(xhr.responseText.includes("testcookie=value123"), "Cookie was sent to upgraded HTTPS request");
      SimpleTest.finish();
    };

    xhr.onerror = function() {
      ok(false, "XHR request failed");
      SimpleTest.finish();
    };

    xhr.send();
  </script>
</body>
</html>
